refactor(event-runtime-web): centralize schedules API calls into api.ts client (OPS-501)
